Reorganizing files in HAL
