+++ .gitignore
@@ -13,6 +13,7 @@ tools/Telemetry
 tools/gcc-arm-none-eabi-*
 tools/clang*
 tools/openocd*
+tools/gnu-mcu-eclipse-openocd*
 
 **/*.gch
 **/*.gcno
