+++ firmware/HelloWorld/source/main.cpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 #include <iterator>
 
-#include "L2_HAL/onboard_led.hpp"
+#include "L2_HAL/displays/led/onboard_led.hpp"
 #include "utility/log.hpp"
 #include "utility/time.hpp"
