+++ firmware/library/L2_HAL/displays/led/test/onboard_led_test.cpp
@@ -1,4 +1,4 @@
-#include "L2_HAL/onboard_led.hpp"
+#include "L2_HAL/displays/led/onboard_led.hpp"
 #include "L4_Testing/testing_frameworks.hpp"
 
 EMIT_ALL_METHODS(OnBoardLed);
