+++ firmware/library/L2_HAL/displays/oled/test/ssd1306_test.cpp
@@ -1,5 +1,5 @@
 // Tests for the St7066u Parallel LCD Driver class.
-#include "L2_HAL/ssd1306.hpp"
+#include "L2_HAL/displays/oled/ssd1306.hpp"
 #include "L4_Testing/testing_frameworks.hpp"
 
 EMIT_ALL_METHODS(Ssd1306);
