+++ firmware/library/L2_HAL/sensors/temperature/temperature.hpp
