+++ firmware/library/L2_HAL/sensors/temperature/test/temperature_test.cpp
@@ -1,4 +1,4 @@
-#include "L2_HAL/temperature.hpp"
+#include "L2_HAL/sensors/temperature/temperature.hpp"
 #include "L4_Testing/testing_frameworks.hpp"
 
 // Uncomment this when can class has been created
